+++ java/resources/org.jetbrains.research.testgenie.java.xml
@@ -1,0 +1,5 @@
+<idea-plugin package="org.jetbrains.research.testgenie.kotlin">
+    <dependencies>
+        <plugin id="org.jetbrains.kotlin"/>
+    </dependencies>
+</idea-plugin>
