+++ kotlin/src/org/jetbrains/research/testspark/kotlin/KotlinPsiHelper.kt
@@ -1,27 +1,33 @@
-package org.jetbrains.research.testspark.helpers.psi.java
+package org.jetbrains.research.testspark.helpers.psi.kotlin
 
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.psi.PsiFile
+import org.jetbrains.research.testspark.core.psi.Language
+import org.jetbrains.research.testspark.core.psi.PsiHelper
 import com.intellij.openapi.diagnostic.Logger
-import com.intellij.openapi.editor.Caret
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.TextRange
-import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiDocumentManager
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiFile
-import com.intellij.psi.PsiMethod
-import com.intellij.psi.util.PsiTreeUtil
-import com.intellij.psi.util.PsiTypesUtil
-import org.jetbrains.research.testspark.helpers.psi.Language
-import org.jetbrains.research.testspark.helpers.psi.PsiClassWrapper
-import org.jetbrains.research.testspark.helpers.psi.PsiHelper
-import org.jetbrains.research.testspark.helpers.psi.PsiMethodWrapper
+import org.jetbrains.kotlin.psi.KtFunction
+import com.intellij.openapi.editor.Caret
+import com.intellij.psi.PsiClass
+import com.intellij.psi.util.parentOfType
+import org.jetbrains.kotlin.descriptors.ClassDescriptor
+import org.jetbrains.kotlin.idea.caches.resolve.analyze
+import org.jetbrains.kotlin.psi.KtClass
+import org.jetbrains.kotlin.psi.KtClassOrObject
+import org.jetbrains.kotlin.psi.KtTypeReference
+import org.jetbrains.kotlin.resolve.BindingContext
+import org.jetbrains.kotlin.resolve.DescriptorToSourceUtils
+import org.jetbrains.kotlin.resolve.lazy.BodyResolveMode
+import org.jetbrains.research.testspark.core.psi.PsiClassWrapper
+import org.jetbrains.research.testspark.core.psi.PsiMethodWrapper
 import org.jetbrains.research.testspark.tools.llm.SettingsArguments
 
-class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
+class KotlinPsiHelper(private val psiFile: PsiFile) : PsiHelper {
 
-    override val language: Language get() = Language.Java
+    override val language: Language get() = Language.Kotlin
 
     private val log = Logger.getInstance(this::class.java)
 
@@ -32,35 +38,32 @@ class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
     }
 
     override fun getSurroundingClass(caretOffset: Int): PsiClassWrapper? {
-        val classElements = PsiTreeUtil.findChildrenOfAnyType(psiFile, PsiClass::class.java)
-        for (cls in classElements) {
-            if (cls.containsOffset(caretOffset)) {
-                val javaClassWrapper = JavaPsiClassWrapper(cls)
-                if (javaClassWrapper.isTestableClass()) {
-                    log.info("Surrounding class for caret in $caretOffset is ${javaClassWrapper.qualifiedName}")
-                    return javaClassWrapper
-                }
+        val element = psiFile.findElementAt(caretOffset)
+        val cls = element?.parentOfType<KtClassOrObject>(withSelf = true)
+
+        if (cls != null) {
+            if (cls.name != null) {
+                val kotlinClassWrapper = KotlinPsiClassWrapper(cls)
+                log.info("Surrounding class for caret in $caretOffset is ${kotlinClassWrapper.qualifiedName}")
+                return kotlinClassWrapper
             }
         }
+
         log.info("No surrounding class for caret in $caretOffset")
         return null
     }
 
     override fun getSurroundingMethod(caretOffset: Int): PsiMethodWrapper? {
-        val methodElements = PsiTreeUtil.findChildrenOfAnyType(psiFile, PsiMethod::class.java)
-        for (method in methodElements) {
-            if (method.body != null && method.containsOffset(caretOffset)) {
-                val surroundingClass =
-                    PsiTreeUtil.getParentOfType(method, PsiClass::class.java) ?: continue
-                val surroundingClassWrapper = JavaPsiClassWrapper(surroundingClass)
-                if (surroundingClassWrapper.isTestableClass()) {
-                    val javaMethod = JavaPsiMethodWrapper(method)
-                    log.info("Surrounding method for caret in $caretOffset is ${javaMethod.methodDescriptor}")
-                    return javaMethod
-                }
-            }
+        val element = psiFile.findElementAt(caretOffset)
+        val method = element?.parentOfType<KtFunction>(withSelf = true)
+
+        if (method != null) {
+            val wrappedMethod = KotlinPsiMethodWrapper(method)
+            log.info("Surrounding method for caret at $caretOffset is ${wrappedMethod.methodDescriptor}")
+            return wrappedMethod
         }
-        log.info("No surrounding method for caret in $caretOffset")
+
+        log.info("No surrounding method for caret at $caretOffset")
         return null
     }
 
@@ -72,13 +75,11 @@ class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
             doc.getText(TextRange(doc.getLineStartOffset(selectedLine), doc.getLineEndOffset(selectedLine)))
 
         if (selectedLineText.isBlank()) {
-            log.info("Line $selectedLine at caret $caretOffset is blank")
+            log.info("Line $selectedLine at caret $caretOffset is not valid")
             return null
         }
         log.info("Surrounding line at caret $caretOffset is $selectedLine")
-
-        // increase by one is necessary due to different start of numbering
-        return selectedLine + 1
+        return selectedLine
     }
 
     override fun collectClassesToTest(
@@ -86,9 +87,7 @@ class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
         classesToTest: MutableList<PsiClassWrapper>,
         caretOffset: Int,
     ) {
-        // check if cut has any none java super class
         val maxPolymorphismDepth = SettingsArguments(project).maxPolyDepth(0)
-
         val cutPsiClass = getSurroundingClass(caretOffset)!!
         var currentPsiClass = cutPsiClass
         for (index in 0 until maxPolymorphismDepth) {
@@ -97,7 +96,7 @@ class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
             }
 
             if (currentPsiClass.superClass == null ||
-                currentPsiClass.superClass!!.qualifiedName.startsWith("java.")
+                currentPsiClass.superClass!!.qualifiedName.startsWith("kotlin.")
             ) {
                 break
             }
@@ -111,21 +110,23 @@ class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
         classesToTest: List<PsiClassWrapper>,
         polyDepthReducing: Int,
     ): MutableSet<PsiClassWrapper> {
-        val interestingPsiClasses: MutableSet<JavaPsiClassWrapper> = mutableSetOf()
+        val interestingPsiClasses: MutableSet<KotlinPsiClassWrapper> = mutableSetOf()
 
-        var currentLevelClasses =
-            mutableListOf<PsiClassWrapper>().apply { addAll(classesToTest) }
+        var currentLevelClasses = mutableListOf<PsiClassWrapper>().apply { addAll(classesToTest) }
 
         repeat(SettingsArguments(project).maxInputParamsDepth(polyDepthReducing)) {
-            val tempListOfClasses = mutableSetOf<JavaPsiClassWrapper>()
+            val tempListOfClasses = mutableSetOf<KotlinPsiClassWrapper>()
 
             currentLevelClasses.forEach { classIt ->
                 classIt.methods.forEach { methodIt ->
-                    (methodIt as JavaPsiMethodWrapper).parameterList.parameters.forEach { paramIt ->
-                        PsiTypesUtil.getPsiClass(paramIt.type)?.let { typeIt ->
-                            JavaPsiClassWrapper(typeIt).let {
-                                if (!it.qualifiedName.startsWith("java.")) {
-                                    interestingPsiClasses.add(it)
+                    (methodIt as KotlinPsiMethodWrapper).parameterList?.parameters?.forEach { paramIt ->
+                        val typeRef = paramIt.typeReference
+                        if (typeRef != null) {
+                            resolveClassInType(typeRef)?.let { psiClass ->
+                                KotlinPsiClassWrapper(psiClass as KtClass).let {
+                                    if (!it.qualifiedName.startsWith("kotlin.")) {
+                                        interestingPsiClasses.add(it)
+                                    }
                                 }
                             }
                         }
@@ -153,19 +154,19 @@ class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
         val caret: Caret =
             e.dataContext.getData(CommonDataKeys.CARET)?.caretModel?.primaryCaret ?: return result.toArray()
 
-        val javaPsiClassWrapped = getSurroundingClass(caret.offset) as JavaPsiClassWrapper?
-        val javaPsiMethodWrapped = getSurroundingMethod(caret.offset) as JavaPsiMethodWrapper?
-        val line: Int? = getSurroundingLine(caret.offset)
+        val ktClass = getSurroundingClass(caret.offset)
+        val ktFunction = getSurroundingMethod(caret.offset)
+        val line: Int? = getSurroundingLine(caret.offset)?.plus(1)
 
-        javaPsiClassWrapped?.let { result.add(getClassDisplayName(it)) }
-        javaPsiMethodWrapped?.let { result.add(getMethodDisplayName(it)) }
+        ktClass?.let { result.add(getClassDisplayName(it)) }
+        ktFunction?.let { result.add(getMethodDisplayName(it)) }
         line?.let { result.add(getLineDisplayName(it)) }
 
-        if (javaPsiClassWrapped != null && javaPsiMethodWrapped != null) {
+        if (ktClass != null && ktFunction != null) {
             log.info(
                 "The test can be generated for: \n " +
-                    " 1) Class ${javaPsiClassWrapped.qualifiedName} \n" +
-                    " 2) Method ${javaPsiMethodWrapped.name} \n" +
+                    " 1) Class ${ktClass.qualifiedName} \n" +
+                    " 2) Method ${ktFunction.name} \n" +
                     " 3) Line $line",
             )
         }
@@ -179,18 +180,20 @@ class JavaPsiHelper(private val psiFile: PsiFile) : PsiHelper {
         "<html><b><font color='orange'>${psiClass.classType.representation}</font> ${psiClass.qualifiedName}</b></html>"
 
     override fun getMethodDisplayName(psiMethod: PsiMethodWrapper): String {
-        return if ((psiMethod as JavaPsiMethodWrapper).isDefaultConstructor) {
-            "<html><b><font color='orange'>default constructor</font></b></html>"
-        } else if (psiMethod.isConstructor) {
-            "<html><b><font color='orange'>constructor</font></b></html>"
-        } else if (psiMethod.isMethodDefault) {
-            "<html><b><font color='orange'>default method</font> ${psiMethod.name}</b></html>"
-        } else {
-            "<html><b><font color='orange'>method</font> ${psiMethod.name}</b></html>"
+        psiMethod as KotlinPsiMethodWrapper
+        return when {
+            psiMethod.isTopLevelFunction -> "<html><b><font color='orange'>top-level function</font></b></html>"
+            psiMethod.isSecondaryConstructor -> "<html><b><font color='orange'>secondary constructor</font></b></html>"
+            psiMethod.isPrimaryConstructor -> "<html><b><font color='orange'>constructor</font></b></html>"
+            psiMethod.isDefaultMethod -> "<html><b><font color='orange'>default method</font> ${psiMethod.name}</b></html>"
+            else -> "<html><b><font color='orange'>method</font> ${psiMethod.name}</b></html>"
         }
     }
 
-    private fun PsiElement.containsOffset(caretOffset: Int): Boolean {
-        return (textRange.startOffset <= caretOffset) && (textRange.endOffset >= caretOffset)
+    private fun resolveClassInType(typeReference: KtTypeReference): PsiClass? {
+        val context = typeReference.analyze(BodyResolveMode.PARTIAL)
+        val type = context[BindingContext.TYPE, typeReference] ?: return null
+        val classDescriptor = type.constructor.declarationDescriptor as? ClassDescriptor ?: return null
+        return (DescriptorToSourceUtils.getSourceFromDescriptor(classDescriptor) as? KtClass)?.toLightClass()
     }
 }
